debate_bot/requirements.txt
# Debate Bot Dependencies
aiogram>=3.0.0
aiohttp>=3.8.0
redis>=4.5.0
prometheus-client>=0.16.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.11.0
cryptography>=41.0.0
aiofiles>=23.0.0
